move and fix code style
