+++ paddle/fluid/memory/allocation/allocator_facade.cc
@@ -68,7 +68,7 @@
 
 #ifdef PADDLE_WITH_CUSTOM_DEVICE
 #include "paddle/fluid/memory/allocation/custom_allocator.h"
-#include "paddle/fluid/platform/device/device_wrapper.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #endif
 
 PADDLE_DEFINE_EXPORTED_int64(
