+++ paddle/fluid/memory/allocation/buddy_allocator.cc
@@ -25,8 +25,8 @@ limitations under the License. */
 DECLARE_uint64(reallocate_gpu_memory_in_mb);
 #endif
 
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/fluid/platform/place.h"
+#include "paddle/phi/backends/device_wrapper.h"
 
 namespace paddle {
 namespace memory {
