+++ paddle/fluid/memory/memcpy.cc
@@ -14,9 +14,9 @@ limitations under the License. */
 
 #include "paddle/fluid/memory/memcpy.h"
 
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/profiler/event_tracing.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/common/place.h"
 
 #ifdef PADDLE_WITH_XPU
