+++ paddle/fluid/operators/log_loss_op_xpu.cc
@@ -13,7 +13,7 @@ limitations under the License. */
 #include <memory>
 
 #include "paddle/fluid/framework/op_registry.h"
-#include "paddle/fluid/platform/device/device_wrapper.h"
+#include "paddle/phi/backends/device_wrapper.h"
 namespace paddle {
 namespace operators {
