+++ paddle/fluid/operators/rnn_op_mlu.cc
@@ -12,9 +12,9 @@ limitations under the License. */
 #include "paddle/fluid/framework/op_registry.h"
 #include "paddle/fluid/operators/mlu/mlu_baseop.h"
 #include "paddle/fluid/operators/utils.h"
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/fluid/platform/device/xpu/xpu_header.h"
 #include "paddle/fluid/platform/device_context.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/kernels/funcs/math_function.h"
 
 namespace paddle {
