+++ paddle/fluid/platform/device/xpu/xpu_info.cc
@@ -15,11 +15,11 @@ limitations under the License. */
 #include <string>
 
 #include "gflags/gflags.h"
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/fluid/platform/device/xpu/enforce_xpu.h"
 #include "paddle/fluid/platform/device/xpu/xpu_header.h"
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/place.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/backends/xpu/xpu_info.h"
 
 namespace paddle {
