+++ paddle/phi/backends/callback_manager.cc
@@ -16,8 +16,8 @@
 
 #include <ThreadPool.h>
 
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/phi/backends/device_guard.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/core/enforce.h"
 
 namespace phi {
