+++ paddle/phi/backends/event.cc
@@ -14,8 +14,8 @@
 
 #include "paddle/phi/backends/event.h"
 
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/phi/backends/device_guard.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/backends/stream.h"
 
 namespace phi {
