+++ paddle/phi/backends/stream.cc
@@ -14,8 +14,8 @@
 
 #include "paddle/phi/backends/stream.h"
 
-#include "paddle/fluid/platform/device/device_wrapper.h"
 #include "paddle/phi/backends/device_guard.h"
+#include "paddle/phi/backends/device_wrapper.h"
 #include "paddle/phi/backends/event.h"
 
 namespace phi {
